Make parseAPTConfigLine public
We already have a way to go from a repository to a line, via
`Repository.APTConfigLine`. There is a valid usecase to have the
reverse, go from a line to a Repository: for instance some software will
expect such a line as input/output (Ansible's apt_repository's repo for
instance). Given that it's unlikely that the implementation of it will
see breaking changes, I don't see any reason to keep it private.